Refactored to separate switch/router/common functionality
