+++ common/common.go
@@ -1,0 +1,64 @@
+package common
+
+import (
+	"fmt"
+	"go.bug.st/serial"
+	"log"
+)
+
+func FormatCommand(cmd string) []byte {
+	formattedString := []byte(cmd + "\n")
+	return formattedString
+}
+
+func ReadLine(port serial.Port, buffSize int, debug bool) []byte {
+	line := ReadLines(port, buffSize, 1, debug)
+	return line[0]
+}
+
+func ReadLines(port serial.Port, buffSize int, maxLines int, debug bool) [][]byte {
+	output := make([][]byte, maxLines)
+
+	for i := 0; i < maxLines; i++ {
+		output[i] = make([]byte, buffSize)
+		for {
+			// Reads up to buffSize bytes
+			n, err := port.Read(output[i])
+			if err != nil {
+				log.Fatal(err)
+			}
+			if n == 0 {
+				break
+			}
+			if debug {
+				fmt.Printf("%s", output[i][:n])
+			}
+			if n == '\n' {
+				break
+			}
+		}
+	}
+
+	return output
+}
+
+func TrimNull(bytes []byte) []byte {
+	friendlyLine := make([]byte, 0)
+	if !IsEmpty(bytes) {
+		for _, val := range bytes {
+			if val != 0x00 {
+				friendlyLine = append(friendlyLine, val)
+			}
+		}
+	}
+	return friendlyLine
+}
+
+func IsEmpty(output []byte) bool {
+	for _, outputByte := range output {
+		if outputByte != byte(0) {
+			return false
+		}
+	}
+	return true
+}
